# Optimize position of ReentrantLock to avoid not necessary acquire lock.