perf: Optimize meter creation with double-checked locking

- Move lock acquisition after initial meter check to reduce lock contention
- Only acquire lock when meter needs to be created
- Implement double-checked locking pattern to ensure thread safety
- Improves performance for frequently accessed existing meters

This change reduces unnecessary synchronization overhead by avoiding
lock acquisition when meters already exist in the registry.

Performing 50VUs, 4m, This is the frameGraph for current jPOS vs fix/metric-lock

Fix/metric-lock
![image](https://github.com/user-attachments/assets/5ffc9e29-ce58-45be-91dd-3a2cffd1a580)
```
checks.........................: 50.00%  1448092 out of 2896184
     data_received..................: 859 MB  3.6 MB/s
     data_sent......................: 130 MB  543 kB/s
     http_req_blocked...............: avg=3.14µs  min=0s       med=1µs    max=7.25ms  p(90)=2µs     p(95)=3µs    
     http_req_connecting............: avg=1.41µs  min=0s       med=0s     max=7.22ms  p(90)=0s      p(95)=0s     
     http_req_duration..............: avg=8.21ms  min=520µs    med=7.95ms max=1.7s    p(90)=11.17ms p(95)=12.38ms
       { expected_response:true }...: avg=8.21ms  min=520µs    med=7.95ms max=1.7s    p(90)=11.17ms p(95)=12.38ms
     http_req_failed................: 0.00%   0 out of 1448092
     http_req_receiving.............: avg=17.63µs min=3µs      med=14µs   max=15.77ms p(90)=26µs    p(95)=34µs   
     http_req_sending...............: avg=4.96µs  min=1µs      med=4µs    max=14.7ms  p(90)=8µs     p(95)=10µs   
     http_req_tls_handshaking.......: avg=0s      min=0s       med=0s     max=0s      p(90)=0s      p(95)=0s     
     http_req_waiting...............: avg=8.19ms  min=426µs    med=7.93ms max=1.7s    p(90)=11.14ms p(95)=12.35ms
     http_reqs......................: 1448092 6033.635414/s
     iteration_duration.............: avg=8.28ms  min=657.29µs med=8.01ms max=1.7s    p(90)=11.24ms p(95)=12.45ms
     iterations.....................: 1448092 6033.635414/s
     vus............................: 50      min=50                 max=50
     vus_max........................: 50      min=50                 max=50
```

main
![image](https://github.com/user-attachments/assets/80eb37ef-fbe6-4a92-95ac-04943815f29f)
```
checks.........................: 50.00% 784565 out of 1569130
     data_received..................: 465 MB 1.9 MB/s
     data_sent......................: 71 MB  294 kB/s
     http_req_blocked...............: avg=4.14µs  min=0s       med=2µs     max=11.89ms  p(90)=3µs     p(95)=4µs    
     http_req_connecting............: avg=1.77µs  min=0s       med=0s      max=11.84ms  p(90)=0s      p(95)=0s     
     http_req_duration..............: avg=15.19ms min=458µs    med=14.46ms max=226.5ms  p(90)=22.71ms p(95)=25.67ms
       { expected_response:true }...: avg=15.19ms min=458µs    med=14.46ms max=226.5ms  p(90)=22.71ms p(95)=25.67ms
     http_req_failed................: 0.00%  0 out of 784565
     http_req_receiving.............: avg=21.1µs  min=4µs      med=18µs    max=10.19ms  p(90)=27µs    p(95)=34µs   
     http_req_sending...............: avg=6.39µs  min=1µs      med=5µs     max=11.75ms  p(90)=9µs     p(95)=11µs   
     http_req_tls_handshaking.......: avg=0s      min=0s       med=0s      max=0s       p(90)=0s      p(95)=0s     
     http_req_waiting...............: avg=15.17ms min=438µs    med=14.43ms max=226.48ms p(90)=22.69ms p(95)=25.64ms
     http_reqs......................: 784565 3268.862906/s
     iteration_duration.............: avg=15.28ms min=519.54µs med=14.55ms max=226.57ms p(90)=22.81ms p(95)=25.77ms
     iterations.....................: 784565 3268.862906/s
     vus............................: 50     min=50                max=50
     vus_max........................: 50     min=50                max=50
```